requirements-dev.txt
codecov==2.0.15
coverage==4.5.4
flake8==3.7.7
mypy==0.740
pycodestyle==2.5.0
pylint==2.4.4
pytest==5.2.0
pytest-cov==2.8.1
requests-mock==1.7.0
